moved HkuLibraryHoursFactory.ts from service\hour\model\ to service\hour\ htbkoo/hey-hkul-hours
